Extracting NFTs capabilities to a specific `nft` api
